+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Fixed rare crash when unable to load Mooshroom textures.
 - Fixed lockdown effects not syncing whilst ticking.
 - Fixed infinite lockdown effects being dimmed and prioritised in the HUD.
+- [NEOFORGE] Fixed `lock_effect` criterion trigger not running.
 
 ### Miscellaneous
 - Update mod icon.